+++ packages/sdk/contractkit/src/wrappers/StableToken.test.ts
@@ -25,6 +25,11 @@ testWithGanache('StableToken Wrapper', async (web3) => {
       name: 'Celo Euro',
       symbol: 'cEUR',
     },
+    [StableToken.cREAL]: {
+      stableToken: StableToken.cREAL,
+      name: 'Celo Brazilian Real',
+      symbol: 'cREAL',
+    },
   }
 
   for (const stableTokenInfo of Object.values(stableTokenInfos)) {